initial the fix implementation
