+++ docs/light-mesh.png
