+++ http-sidecar/src/test/java/com/networknt/router/RouterHttpTest.java
@@ -16,102 +16,32 @@
 
 package com.networknt.router;
 
-import com.networknt.TestServer;
 import com.networknt.client.Http2Client;
-import com.networknt.common.DecryptUtil;
-import com.networknt.common.SecretConstants;
-import com.networknt.config.Config;
 import com.networknt.exception.ClientException;
 import com.networknt.httpstring.HttpStringConstants;
-import com.networknt.server.ServerConfig;
-import io.undertow.Undertow;
 import io.undertow.UndertowOptions;
 import io.undertow.client.ClientConnection;
 import io.undertow.client.ClientRequest;
 import io.undertow.client.ClientResponse;
-import io.undertow.server.HttpHandler;
-import io.undertow.server.HttpServerExchange;
 import io.undertow.util.Headers;
 import io.undertow.util.Methods;
 import org.junit.*;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.xnio.IoUtils;
 import org.xnio.OptionMap;
 
-import javax.net.ssl.*;
-import java.io.InputStream;
 import java.net.URI;
-import java.security.KeyStore;
-import java.security.KeyStoreException;
-import java.security.NoSuchAlgorithmException;
-import java.security.UnrecoverableKeyException;
 import java.util.List;
-import java.util.Map;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.atomic.AtomicReference;
 
-import static com.networknt.server.Server.TRUST_ALL_CERTS;
-import static io.undertow.Handlers.path;
-
-public class RouterHttpTest {
-    static final Logger logger = LoggerFactory.getLogger(RouterHttpTest.class);
-    public static final String CONFIG_NAME = "server";
-    public static final String CONFIG_SECRET = "secret";
-
-    static Undertow server1 = null;
-    @ClassRule
-    public static TestServer server = TestServer.getInstance();
-    public static ServerConfig config = (ServerConfig) Config.getInstance().getJsonObjectConfig(CONFIG_NAME, ServerConfig.class);
-    public static Map<String, Object> secret = DecryptUtil.decryptMap(Config.getInstance().getJsonMapConfig(CONFIG_SECRET));
-    static SSLContext sslContext = createSSLContext();
-
-    static final boolean enableHttp2 = server.getServerConfig().isEnableHttp2();
-    static final boolean enableHttps = server.getServerConfig().isEnableHttps();
-    static final int httpPort = server.getServerConfig().getHttpPort();
-    static final int httpsPort = server.getServerConfig().getHttpsPort();
-    static final String url = "https://localhost:8082";
-
-    @BeforeClass
-    public static void setUp() {
-        if(server1 == null) {
-            logger.info("starting server1");
-            Undertow.Builder builder1 = Undertow.builder()
-                    .addHttpsListener(8082, "localhost", sslContext)
-                    .setHandler(path().addPrefixPath("/v2/address", new HttpHandler() {
-                        @Override
-                        public void handleRequest(HttpServerExchange exchange) throws Exception {
-                            exchange.getResponseHeaders().put(Headers.CONTENT_TYPE, "text/plain");
-                            exchange.getResponseSender().send("Server1");
-                        }
-                    }));
-            if(enableHttp2) {
-                builder1.setServerOption(UndertowOptions.ENABLE_HTTP2, true);
-            }
-            server1 = builder1.build();
-            server1.start();
-        }
-
-    }
-
-    @AfterClass
-    public static void tearDown() throws Exception {
-        if(server1 != null) {
-            try {
-                Thread.sleep(100);
-            } catch (InterruptedException ignored) {
-
-            }
-            server1.stop();
-            logger.info("The router server1 is stopped.");
-        }
-    }
+public class RouterHttpTest extends BaseRouterTest{
 
     /**
      * Calling server1 directly to ensure that the endpoint is working.
      * @throws Exception
      */
+    @Ignore
     @Test
     public void testServer1() throws Exception {
         final Http2Client client = Http2Client.getInstance();
@@ -142,6 +72,7 @@ public class RouterHttpTest {
         }
     }
 
+    @Ignore
     @Test
     public void testGet() throws Exception {
         final Http2Client client = Http2Client.getInstance();
@@ -181,85 +112,6 @@ public class RouterHttpTest {
             if(logger.isDebugEnabled()) logger.debug("body = " + body);
             Assert.assertTrue(body.contains("Server"));
         }
-    }
-
-    private static SSLContext createSSLContext() throws RuntimeException {
-        try {
-            KeyManager[] keyManagers = buildKeyManagers(loadKeyStore(), ((String)secret.get(SecretConstants.SERVER_KEY_PASS)).toCharArray());
-            TrustManager[] trustManagers;
-            if(config.isEnableTwoWayTls()) {
-                trustManagers = buildTrustManagers(loadTrustStore());
-            } else {
-                trustManagers = buildTrustManagers(null);
-            }
-
-            SSLContext sslContext;
-            sslContext = SSLContext.getInstance("TLSv1");
-            sslContext.init(keyManagers, trustManagers, null);
-            return sslContext;
-        } catch (Exception e) {
-            logger.error("Unable to create SSLContext", e);
-            throw new RuntimeException("Unable to create SSLContext", e);
-        }
-    }
-
-    private static KeyStore loadKeyStore() {
-        String name = config.getKeystoreName();
-        try (InputStream stream = Config.getInstance().getInputStreamFromFile(name)) {
-            KeyStore loadedKeystore = KeyStore.getInstance("JKS");
-            loadedKeystore.load(stream, ((String)secret.get(SecretConstants.SERVER_KEYSTORE_PASS)).toCharArray());
-            return loadedKeystore;
-        } catch (Exception e) {
-            logger.error("Unable to load keystore " + name, e);
-            throw new RuntimeException("Unable to load keystore " + name, e);
-        }
-    }
-
-    protected static KeyStore loadTrustStore() {
-        String name = config.getTruststoreName();
-        try (InputStream stream = Config.getInstance().getInputStreamFromFile(name)) {
-            KeyStore loadedKeystore = KeyStore.getInstance("JKS");
-            loadedKeystore.load(stream, ((String)secret.get(SecretConstants.SERVER_TRUSTSTORE_PASS)).toCharArray());
-            return loadedKeystore;
-        } catch (Exception e) {
-            logger.error("Unable to load truststore " + name, e);
-            throw new RuntimeException("Unable to load truststore " + name, e);
-        }
-    }
-
-    private static TrustManager[] buildTrustManagers(final KeyStore trustStore) {
-        TrustManager[] trustManagers = null;
-        if (trustStore == null) {
-            try {
-                TrustManagerFactory trustManagerFactory = TrustManagerFactory
-                        .getInstance(KeyManagerFactory.getDefaultAlgorithm());
-                trustManagerFactory.init(trustStore);
-                trustManagers = trustManagerFactory.getTrustManagers();
-            }
-            catch (NoSuchAlgorithmException | KeyStoreException e) {
-                logger.error("Unable to initialise TrustManager[]", e);
-                throw new RuntimeException("Unable to initialise TrustManager[]", e);
-            }
-        }
-        else {
-            trustManagers = TRUST_ALL_CERTS;
-        }
-        return trustManagers;
-    }
-
-    private static KeyManager[] buildKeyManagers(final KeyStore keyStore, char[] keyPass) {
-        KeyManager[] keyManagers;
-        try {
-            KeyManagerFactory keyManagerFactory = KeyManagerFactory.getInstance(KeyManagerFactory
-                    .getDefaultAlgorithm());
-            keyManagerFactory.init(keyStore, keyPass);
-            keyManagers = keyManagerFactory.getKeyManagers();
-        }
-        catch (NoSuchAlgorithmException | UnrecoverableKeyException | KeyStoreException e) {
-            logger.error("Unable to initialise KeyManager[]", e);
-            throw new RuntimeException("Unable to initialise KeyManager[]", e);
-        }
-        return keyManagers;
     }
 
 }
