+++ http-sidecar/src/test/resources/config/handler.yml
@@ -45,6 +45,7 @@ handlers:
   # Header middleware to manipulate request and/or response headers before or after downstream server
   - com.networknt.header.HeaderHandler@header
   # Direct requests to named services based on the request path
+  - com.networknt.router.middleware.SAMLTokenHandler@saml
   - com.networknt.router.middleware.ServiceDictHandler@path
   - com.networknt.router.RouterHandler@router
   - com.networknt.proxy.LightProxyHandler@proxy
@@ -74,6 +75,16 @@ chains:
     #- sanitizer
     - validator
     - proxy
+  router:
+    - exception
+    # - metrics
+    - traceability
+    - correlation
+    # - cors
+    - header
+    # - saml
+    - path
+    - router
 
 paths:
   - path: '/*'
@@ -96,6 +107,10 @@ paths:
     method: 'patch'
     exec:
       - default
+  - path: '/v2/address'
+    method: 'get'
+    exec:
+      - router
 
   - path: '/health/com.networknt.backend-1.0.0'
     method: 'get'
