+++ http-sidecar/src/test/resources/config/proxy.yml
@@ -8,7 +8,8 @@ http2Enabled: false
 httpsEnabled: false
 
 # Target URIs
-hosts: http://localhost:8081
+hosts:
+  - http://localhost:8081
 
 # Connections per thread to the target servers
 connectionsPerThread: 20
