+++ http-sidecar/src/test/resources/config/token.yml
@@ -1,1 +1,2 @@
 enabled: true
+scopeToken: true
